+++ src/composables/Defense/StatCalculations/DefenseRange.ts
@@ -7,6 +7,7 @@ import type { ModInterface, ShardInterface } from '@/types';
 import useAncientPowers from '@/composables/Defense/AncientPowers';
 import usePylonCalculations from '@/composables/Defense/PylonCalculations';
 import useModsShards from '@/composables/Defense/StatCalculations/ModsShards';
+import OutputModifier from '@/classes/OutputModifier';
 
 interface DefenseRangeCalculationsComposable {
     defenseRange: ComputedRef<number>
@@ -29,11 +30,11 @@ export default function useDefenseRangeCalculations(
         let rangeAdditive = defense.userData.relic.godlyStat?.type === 'defense_range' ? defense.userData.relic.godlyStat.value : 0;
         let rangeMultiplier = 1;
 
-        forRegularModsAndShards('defenseRange', (util: ModInterface | ShardInterface) => {
-            rangeAdditive += util.defenseRange?.additive ?? 0
+        forRegularModsAndShards('defenseRange', (util: ModInterface | ShardInterface, defenseRangeModifier: OutputModifier) => {
+            rangeAdditive += defenseRangeModifier.additive ?? 0
             
-            if (util.defenseRange?.percentage) {
-                rangeMultiplier *= (1 + util.defenseRange.percentage / 100)
+            if (defenseRangeModifier.percentage) {
+                rangeMultiplier *= (1 + defenseRangeModifier.percentage / 100)
             }
         })