+++ src/composables/Defense/StatCalculations/DefenseRate.ts
@@ -6,6 +6,7 @@ import type { CalculationConditionsInterface } from '@/composables/Defense/Defen
 import type { ModInterface, ShardInterface } from '@/types';
 import usePylonCalculations from '@/composables/Defense/PylonCalculations';
 import useModsShards from '@/composables/Defense/StatCalculations/ModsShards';
+import OutputModifier from '@/classes/OutputModifier';
 
 interface DefenseRateCalculationsComposable {
     attackRate: ComputedRef<number>
@@ -24,8 +25,8 @@ export default function useDefenseRateCalculations(
 
         let attackRatePercentage: number = 0;
         
-        forRegularModsAndShards('defenseRate', (util: ModInterface | ShardInterface) => {
-            attackRatePercentage += util.defenseRate?.percentage ?? 0
+        forRegularModsAndShards('defenseRate', (util: ModInterface | ShardInterface, defenseRateModifier: OutputModifier) => {
+            attackRatePercentage += defenseRateModifier.percentage ?? 0
         })
         
         attackRatePercentage += defense.ascensionRate;